Unnest if statements
Overall, it is generally considered bad style to continually nest loops. However, if this is what the decompiler outputted, then why bother touching it?

However, this can be rewritten as returning early if the playtime is less than 60.

The reason for this is not because of the original code, but because it looks better, has the same output, and is easier to understand.